Add kernel and initrd info to preseed template
This adds kernel_name, kernel_description, initrd_name, and
initrd_description as jinja2 variables to the preseed template.

- Add test cases for preseed jinja2 resolution
  - Create tests/preseed for testing get_preseed()
  - Move test setup unrelated to api from tests/api/conftest to
    tests/conftest so that it can be re-used in tests/preseed
